Minor file reorganisation
